ci: prune extra threaded builds
Swish may run in threaded Chez Scheme for convenience, but this is not a
primary supported target. We include threaded builds to help us avoid
breaking changes, but we can scale this back to Windows and Linux on
latest stable Chez Scheme and limit variety of OS releases. GitHub's
macOS runners are too slow to include in this screening.