profile: fix destination for annotated files
If a source file does (include "../../aux/file.ss"), then swish-test
could end up trying to write an annotated file.ss.html outside the
target directory containing the specified coverage summary report.

To fix this we determine a common prefix based on the real paths of the
profiled source files and use the distinct suffix as the path within the
target directory.

The appearance of the coverage report changes slightly since we:
 - sort entries based on the real path of the file,
 - use natural-string<? for this sort, and
 - strip the common prefix from filenames.

 The result is often more concise than the original and helps to avoid
 broken hyperlinks in the summary report.