support procedural process-trap-exit handler
Support a procedural handler as the value for process-trap-exit
so that a process blocked in a gen-server call can perform some
clean-up if killed.

This commit also fixes an obscure bug in the case where a receive
form has a guard that calls receive.

Rework the interactive console so that:
 - we can interrupt the REPL via ^C on Windows again,
 - we can wait until interrupts are enabled before invoking the
   keyboard-interrupt-handler, and
 - a script can start an interactive REPL that handles ^C as Swish does

User visible changes to the behavior of the interactive console include:
 - trying to read from the console in a spawned process returns #!eof,