fix broken http redirect for /
The recent change to path-combine means that (path-combine "abc" "")
now returns "abc" instead of "abc/". It looks like the http server had
relied on that and may also have been using substring to work around the
case where (path-combine "a" "/b") used to return "a//b".
The simplification here fixes redirect for /.